moved writeMetadata to VSM
